pim-compiler/output/react_agent_enhanced_v3/requirements.txt
```text
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==1.4.23
alembic==1.7.5
psycopg2-binary==2.9.1
python-multipart==0.0.5
email-validator==1.1.3
```